Rewrite API using lambdas instead of traits
