+++ examples/02_structs.rs
@@ -1,15 +1,12 @@
 use anyhow::Result;
-use async_trait::async_trait;
 use env_logger::{Builder, Env};
 use tokio::time::{sleep, Duration};
-use tokio_graceful_shutdown::{AsyncSubsystem, SubsystemHandle, Toplevel};
+use tokio_graceful_shutdown::{SubsystemHandle, Toplevel};
 
 struct Subsystem1 {}
 
-#[async_trait]
-impl AsyncSubsystem for Subsystem1 {
-    async fn run(mut self, mut subsys: SubsystemHandle) -> Result<()> {
-        subsys.start("Subsys2", Subsystem2 {});
+impl Subsystem1 {
+    async fn run(self, subsys: SubsystemHandle) -> Result<()> {
         log::info!("Subsystem1 started.");
         subsys.on_shutdown_requested().await;
         log::info!("Shutting down Subsystem1 ...");
@@ -18,26 +15,17 @@ impl AsyncSubsystem for Subsystem1 {
         Ok(())
     }
 }
-struct Subsystem2 {}
-
-#[async_trait]
-impl AsyncSubsystem for Subsystem2 {
-    async fn run(mut self, _subsys: SubsystemHandle) -> Result<()> {
-        log::info!("Subsystem2 started.");
-        sleep(Duration::from_millis(500)).await;
-
-        panic!("Subsystem2 panicked!")
-    }
-}
 
 #[tokio::main]
 async fn main() -> Result<()> {
     // Init logging
     Builder::from_env(Env::default().default_filter_or("debug")).init();
 
+    let subsys = Subsystem1 {};
+
     // Create toplevel
     Toplevel::new()
-        .start("Subsys1", Subsystem1 {})
+        .start("Subsys1", |a| subsys.run(a))
         .catch_signals()
         .wait_for_shutdown(Duration::from_millis(1000))
         .await
