+++ examples/03_shutdown_timeout.rs
@@ -7,7 +7,7 @@ async fn subsys1(subsys: SubsystemHandle) -> Result<()> {
     log::info!("Subsystem1 started.");
     subsys.on_shutdown_requested().await;
     log::info!("Shutting down Subsystem1 ...");
-    sleep(Duration::from_millis(500)).await;
+    sleep(Duration::from_millis(1000)).await;
     log::info!("Subsystem1 stopped.");
     Ok(())
 }
@@ -21,6 +21,6 @@ async fn main() -> Result<()> {
     Toplevel::new()
         .start("Subsys1", subsys1)
         .catch_signals()
-        .wait_for_shutdown(Duration::from_millis(1000))
+        .wait_for_shutdown(Duration::from_millis(500))
         .await
 }
