+++ examples/04_subsystem_finished.rs
@@ -3,12 +3,12 @@ use env_logger::{Builder, Env};
 use tokio::time::{sleep, Duration};
 use tokio_graceful_shutdown::{SubsystemHandle, Toplevel};
 
-async fn subsys1(subsys: SubsystemHandle) -> Result<()> {
+async fn subsys1(_subsys: SubsystemHandle) -> Result<()> {
     log::info!("Subsystem1 started.");
-    subsys.on_shutdown_requested().await;
-    log::info!("Shutting down Subsystem1 ...");
     sleep(Duration::from_millis(500)).await;
     log::info!("Subsystem1 stopped.");
+
+    // Task ends without an error. This should not cause the main program to shutdown.
     Ok(())
 }
