+++ examples/05_subsystem_finished_with_error.rs
@@ -1,15 +1,15 @@
-use anyhow::Result;
+use anyhow::{anyhow, Result};
 use env_logger::{Builder, Env};
 use tokio::time::{sleep, Duration};
 use tokio_graceful_shutdown::{SubsystemHandle, Toplevel};
 
-async fn subsys1(subsys: SubsystemHandle) -> Result<()> {
+async fn subsys1(_subsys: SubsystemHandle) -> Result<()> {
     log::info!("Subsystem1 started.");
-    subsys.on_shutdown_requested().await;
-    log::info!("Shutting down Subsystem1 ...");
     sleep(Duration::from_millis(500)).await;
     log::info!("Subsystem1 stopped.");
-    Ok(())
+
+    // Task ends with an error. This should cause the main program to shutdown.
+    Err(anyhow!("Subsystem1 threw an error."))
 }
 
 #[tokio::main]
