+++ examples/06_nested_subsystems.rs
@@ -3,12 +3,22 @@ use env_logger::{Builder, Env};
 use tokio::time::{sleep, Duration};
 use tokio_graceful_shutdown::{SubsystemHandle, Toplevel};
 
-async fn subsys1(subsys: SubsystemHandle) -> Result<()> {
+async fn subsys1(mut subsys: SubsystemHandle) -> Result<()> {
+    subsys.start("Subsys2", subsys2);
     log::info!("Subsystem1 started.");
     subsys.on_shutdown_requested().await;
     log::info!("Shutting down Subsystem1 ...");
     sleep(Duration::from_millis(500)).await;
     log::info!("Subsystem1 stopped.");
+    Ok(())
+}
+
+async fn subsys2(subsys: SubsystemHandle) -> Result<()> {
+    log::info!("Subsystem2 started.");
+    subsys.on_shutdown_requested().await;
+    log::info!("Shutting down Subsystem2 ...");
+    sleep(Duration::from_millis(500)).await;
+    log::info!("Subsystem2 stopped.");
     Ok(())
 }
