+++ examples/09_task_cancellation.rs
@@ -1,8 +1,14 @@
+//! This example demonstrates how to implement a clean shutdown
+//! of a subsystem.
+//!
+//! The central mechanism here is tokio::select, which can cancel
+//! tasks once one of them finishes. This can be used to cancel
+//! tasks once the shutdown was initiated.
+
 use anyhow::Result;
-use async_trait::async_trait;
 use env_logger::{Builder, Env};
 use tokio::time::{sleep, Duration};
-use tokio_graceful_shutdown::{AsyncSubsystem, SubsystemHandle, Toplevel};
+use tokio_graceful_shutdown::{SubsystemHandle, Toplevel};
 
 struct CountdownSubsystem {}
 impl CountdownSubsystem {
@@ -12,21 +18,20 @@ impl CountdownSubsystem {
 
     async fn countdown(&self) {
         for i in (1..10).rev() {
-            log::info!("Shutting down in: {}", i);
+            log::info!("Countdown: {}", i);
             sleep(Duration::from_millis(1000)).await;
         }
     }
-}
 
-#[async_trait]
-impl AsyncSubsystem for CountdownSubsystem {
-    async fn run(mut self, subsys: SubsystemHandle) -> Result<()> {
+    async fn run(self, subsys: SubsystemHandle) -> Result<()> {
+        log::info!("Starting countdown ...");
+
         tokio::select! {
             _ = subsys.on_shutdown_requested() => {
                 log::info!("Countdown cancelled.");
             },
             _ = self.countdown() => {
-                subsys.request_shutdown();
+                log::info!("Countdown finished.");
             }
         };
 
@@ -41,7 +46,7 @@ async fn main() -> Result<()> {
 
     // Create toplevel
     Toplevel::new()
-        .start("Countdown", CountdownSubsystem::new())
+        .start("Countdown", |h| CountdownSubsystem::new().run(h))
         .catch_signals()
         .wait_for_shutdown(Duration::from_millis(1000))
         .await
