--- src/subsys/mod.rs
@@ -1,3 +1,0 @@
-mod lambda;
-
-pub use lambda::LambdaSubsystem;
